backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
opencv-python==4.10.0.84
torch==2.6.0
torchvision==0.21.0
insightface==0.7.3
onnxruntime==1.20.1
scikit-learn>=1.6.0
hdbscan==0.8.41
faiss-cpu==1.9.0.post1
numpy==1.26.4
Pillow==10.4.0
aiosqlite==0.20.0
bcrypt==4.1.2
